% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_to_json.R
\name{xlsx_to_json}
\alias{xlsx_to_json}
\title{xlsx_to_json}
\usage{
xlsx_to_json(base64_xlsx)
}
\arguments{
\item{base64_xlsx}{base64 of excel workbook.}
}
\value{
JSON of individual worksheets.
}
\description{
Accepts base64 string of excel workbook and outputs a JSON with base64 of individual worksheets.
}
