% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{xlsx_to_json}
\alias{xlsx_to_json}
\title{xlsx_to_json

This functions accepts the base 64 encoded string of an excel workbook, and outputs a JSON with base 64 strings of the worksheets contained within the workbook.
@param base64_xlsx base64 string of the excel workbook}
\usage{
xlsx_to_json(base64_xlsx)
}
\description{
xlsx_to_json

This functions accepts the base 64 encoded string of an excel workbook, and outputs a JSON with base 64 strings of the worksheets contained within the workbook.
@param base64_xlsx base64 string of the excel workbook
}
